[NEAT-390] ⛓Conversion methods as Rule Transformations Part 2 (#588)
* refactor: remove as_information_rules

* refactor: removed unused

* refactor: remove info.as_dms_rules

* refactor: remove asset.as_dms_rules

* build: changelog

* refactor: removed unused
