+++ tests/tests_unit/rules/test_models/test_dms_rules.py
@@ -24,6 +24,7 @@ from cognite.neat.rules.models.dms import (
     DMSSchema,
     DMSViewInput,
 )
+from cognite.neat.rules.transformers import DMSToInformation, InformationToDMS
 from cognite.neat.utils.cdf.data_classes import ContainerApplyDict, NodeApplyDict, SpaceApplyDict, ViewApplyDict
 from tests.data import car
 
@@ -1458,7 +1459,7 @@ class TestDMSRules:
 
     def test_alice_as_information(self, alice_spreadsheet: dict[str, dict[str, Any]]) -> None:
         alice_rules = DMSRulesInput.load(alice_spreadsheet).as_rules()
-        info_rules = alice_rules.as_information_rules()
+        info_rules = DMSToInformation().transform(alice_rules)
 
         assert isinstance(info_rules, InformationRules)
 
@@ -1528,7 +1529,7 @@ class TestDMSRules:
         for view in info_rules_copy.views:
             view.reference = None
 
-        info_rules = olav_dms_rules.as_information_rules()
+        info_rules = DMSToInformation().transform(olav_dms_rules)
 
         assert isinstance(info_rules, InformationRules)
 
@@ -1563,7 +1564,7 @@ class TestDMSRules:
         assert sorted(actual_issues) == sorted(expected_issues)
 
     def test_create_reference(self) -> None:
-        dms_rules = car.CAR_RULES.as_dms_rules()
+        dms_rules = InformationToDMS().transform(car.CAR_RULES)
 
         dms_rules.create_reference(car.BASE_MODEL, {"Manufacturer": "Entity", "Color": "Entity"})
 
@@ -1695,7 +1696,7 @@ class TestDMSExporter:
         assert not missing_properties, f"Missing properties for views: {missing_properties}"
 
     def test_camilla_business_solution_as_schema(self, camilla_information_rules: InformationRules) -> None:
-        dms_rules = camilla_information_rules.as_dms_rules()
+        dms_rules = InformationToDMS().transform(camilla_information_rules)
         expected_views = {"TimeseriesForecastProduct", "WindFarm"}
 
         schema = dms_rules.as_schema()
